codec: SMFIR_REPLBODY's body string is not necessarily zero-terminated
The (reverse engineered) spec I've seen says that the 'body' string
in a SMFIR_REPLBODY messages from a milter client is zero-terminated.
PureMessage begs to disagree; it sends non-zero-terminated REPLBODY
messages (and it sends more than one of them). In a dispute between
the spec and real implementations, real implementations win.

This requires a new field type just for SMFIR_REPLBODY (and some
tests to go with it), but that's life.

I've also puts a huge lot more information in the 'incomplete message'
exception raised, because it turned out to be necessary to debug this
issue. It's all as a string right now because this is a fast hack and
I want to think about how it should be structured if it is going to be
structured.