codec & convo: correctly handle the milter side of option negotiation
My previous code was laughably stupid about this, because I had missed
that the 'protocol' field wanted special handling because it is a
negative field, not a positive field: it is not protocol elements that
the milter wants the MTA to send, it is protocol elements that it wants
the MTA to skip. This means that the milter side cannot treat it the
same as the MTA side.