codec: make encode_optneg() be actually useful
I'm not sure it's the right answer. Proper usage requires you to
combine optneg_mta_capable() or optneg_milter_capable() and then
call encode_optneg() the right way, and this feels like a crappy
interface that we should roll into one, but then it takes a lot
of arguments and returns a lot of results.

API design is hard, let's go coding.